requirements.txt
streamlit
pandas
yfinance
plotly
PyPDF2
google-genai